Reorganize python module to be somewhat easier to maintain
